Fixed some compiler warnings.  Missing casts from double to int, size_t type in printf/Form statements, undefined variables.  Nothing major...
git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3071 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Parity/src/QwSubsystemArrayParity.cc
@@ -445,7 +445,7 @@ void  QwSubsystemArrayParity::ConstructBranchAndVector(TTree *tree, TString& pre
 void QwSubsystemArrayParity::FillTreeVector(std::vector<Double_t>& values) const
 {
   QwSubsystemArray::FillTreeVector(values);
-  if (fErrorFlagTreeIndex>=0 && fErrorFlagTreeIndex<values.size()){
+  if (fErrorFlagTreeIndex>=0 && fErrorFlagTreeIndex<static_cast<int>(values.size())){
     values.at(fErrorFlagTreeIndex) = fErrorFlag;
   }
 }
